split context_utils
