+++ test/CMakeLists.txt
@@ -28,8 +28,7 @@ target_sources(
     boost_mysql_testing
     INTERFACE
     ${CMAKE_CURRENT_SOURCE_DIR}/common/src/entry_point.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/common/src/tracker_executor.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/common/src/printing.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/common/src/utils.cpp
 )
 target_link_libraries(
     boost_mysql_testing
