+++ test/common/include/test_common/network_result.hpp
@@ -35,8 +35,8 @@
 #include <utility>
 #include <vector>
 
-#include "test_common/context_utils.hpp"
 #include "test_common/create_diagnostics.hpp"
+#include "test_common/poll_until.hpp"
 #include "test_common/printing.hpp"
 #include "test_common/source_location.hpp"
 #include "test_common/tracker_executor.hpp"
