+++ test/common/include/test_common/poll_until.hpp
@@ -5,8 +5,8 @@
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
 
-#ifndef BOOST_MYSQL_TEST_COMMON_INCLUDE_TEST_COMMON_CONTEXT_UTILS_HPP
-#define BOOST_MYSQL_TEST_COMMON_INCLUDE_TEST_COMMON_CONTEXT_UTILS_HPP
+#ifndef BOOST_MYSQL_TEST_COMMON_INCLUDE_TEST_COMMON_POLL_UNTIL_HPP
+#define BOOST_MYSQL_TEST_COMMON_INCLUDE_TEST_COMMON_POLL_UNTIL_HPP
 
 #include <boost/asio/any_io_executor.hpp>
 #include <boost/asio/io_context.hpp>
@@ -19,7 +19,6 @@ namespace boost {
 namespace mysql {
 namespace test {
 
-// TODO: split
 // TODO: remove the executor overloads
 // poll until *done == true
 void poll_until(asio::io_context& ctx, const bool* done, source_location loc = BOOST_MYSQL_CURRENT_LOCATION);
@@ -40,14 +39,6 @@ void poll_until(
     const std::function<bool()>& done,
     source_location loc = BOOST_MYSQL_CURRENT_LOCATION
 );
-
-struct io_context_fixture
-{
-    asio::io_context ctx;
-
-    // Checks that we effectively run out of work
-    ~io_context_fixture();
-};
 
 }  // namespace test
 }  // namespace mysql
