+++ test/integration/include/test_integration/run_coro.hpp
@@ -16,7 +16,7 @@
 
 #include <functional>
 
-#include "test_common/context_utils.hpp"
+#include "test_common/poll_until.hpp"
 #include "test_common/source_location.hpp"
 
 namespace boost {
