+++ test/integration/test/connection_pool.cpp
@@ -34,8 +34,8 @@
 #include <utility>
 
 #include "test_common/ci_server.hpp"
-#include "test_common/context_utils.hpp"
 #include "test_common/create_basic.hpp"
+#include "test_common/io_context_fixture.hpp"
 #include "test_common/network_result.hpp"
 #include "test_common/printing.hpp"
 #include "test_common/tracker_executor.hpp"
