+++ test/unit/test/connection_pool/connection_pool_impl.cpp
@@ -54,9 +54,10 @@
 #include <ostream>
 #include <utility>
 
-#include "test_common/context_utils.hpp"
 #include "test_common/create_diagnostics.hpp"
+#include "test_common/io_context_fixture.hpp"
 #include "test_common/network_result.hpp"
+#include "test_common/poll_until.hpp"
 #include "test_common/printing.hpp"
 #include "test_common/source_location.hpp"
 #include "test_common/tracker_executor.hpp"
