+++ test/unit/test/connection_pool/run_with_timeout.cpp
@@ -32,9 +32,9 @@
 #include <chrono>
 #include <utility>
 
-#include "test_common/context_utils.hpp"
+#include "test_common/io_context_fixture.hpp"
+#include "test_common/poll_until.hpp"
 #include "test_common/printing.hpp"
-#include "test_common/tracker_executor.hpp"
 #include "test_unit/mock_timer.hpp"
 
 #ifdef __cpp_lib_polymorphic_allocator
