+++ test/unit/test/detail/engine_impl.cpp
@@ -28,11 +28,10 @@
 #include <cstdint>
 #include <cstring>
 
-#include "test_common/context_utils.hpp"
 #include "test_common/create_diagnostics.hpp"
+#include "test_common/io_context_fixture.hpp"
 #include "test_common/netfun_maker.hpp"
 #include "test_common/network_result.hpp"
-#include "test_common/tracker_executor.hpp"
 #include "test_unit/printing.hpp"
 
 using namespace boost::mysql::test;
