+++ test/unit/test/spotchecks/multifn.cpp
@@ -16,7 +16,7 @@
 #include <boost/test/unit_test.hpp>
 
 #include "test_common/check_meta.hpp"
-#include "test_common/context_utils.hpp"
+#include "test_common/io_context_fixture.hpp"
 #include "test_common/network_result.hpp"
 #include "test_unit/create_coldef_frame.hpp"
 #include "test_unit/create_frame.hpp"
