+++ test/unit/test/with_diagnostics.cpp
@@ -23,8 +23,9 @@
 #include <memory>
 #include <type_traits>
 
-#include "test_common/context_utils.hpp"
 #include "test_common/create_diagnostics.hpp"
+#include "test_common/io_context_fixture.hpp"
+#include "test_common/poll_until.hpp"
 #include "test_common/printing.hpp"
 #include "test_common/tracker_executor.hpp"
 #include "test_unit/create_err.hpp"
